Move user settings to account app
Also fixes the active menu item task taiga#844
