+++ apps/account/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class Config(AppConfig):
+    name = 'apps.account'
+    label = 'meinberlin_account'
