+++ apps/account/forms.py
@@ -1,0 +1,10 @@
+from django import forms
+
+from apps.users.models import User
+
+
+class ProfileForm(forms.ModelForm):
+
+    class Meta:
+        model = User
+        fields = ['username', ]
