+++ apps/account/templates/meinberlin_account/password.html
@@ -1,4 +1,4 @@
-{% extends "meinberlin_dashboard/base_user.html" %}
+{% extends "meinberlin_account/account_dashboard.html" %}
 
 {% load i18n %}
 
@@ -9,7 +9,7 @@
 
     <div class="general-form">
         <div class="row">
-            <form name="PasswordChangeForm" action="{% url 'dashboard-password' %}" method="post" class="l-col-2">
+            <form name="PasswordChangeForm" action="{% url 'account-password' %}" method="post" class="l-col-2">
                 {% csrf_token %}
 
                 {% if form.non_field_errors %}
