+++ apps/account/templates/meinberlin_account/profile.html
@@ -1,4 +1,4 @@
-{% extends "meinberlin_dashboard/base_user.html" %}
+{% extends "meinberlin_account/account_dashboard.html" %}
 
 {% load i18n %}
