+++ apps/dashboard/templates/meinberlin_dashboard/base_dashboard.html
@@ -43,7 +43,7 @@
                 <ul class="dashboard-nav__pages">
                     <li class="dashboard-nav__page">
                         <a href="{% url 'dashboard-project-list' organisation_slug=view.organisation.slug %}"
-                           class="dashboard-nav__page-link {% ifequal url_name 'dashboard-project-list' %}is-active{% endifequal %}">
+                           class="dashboard-nav__page-link {% ifequal view.menu_item 'project' %}is-active{% endifequal %}">
                             <i class="fa fa-file-o"></i>
                             {% trans 'Participation processes' %}
                         </a>
@@ -51,7 +51,7 @@
 
                     <li class="dashboard-nav__page">
                         <a href="{% url 'dashboard-organisation-edit' organisation_slug=view.organisation.slug %}"
-                           class="dashboard-nav__page-link {% ifequal url_name 'dashboard-organisation-edit' %}is-active{% endifequal %}">
+                           class="dashboard-nav__page-link {% ifequal view.menu_item 'organisation' %}is-active{% endifequal %}">
                             <i class="fa fa-cog"></i>
                             {% trans 'Settings' %}
                         </a>
