+++ apps/dashboard/views.py
@@ -1,7 +1,5 @@
-from allauth.account import views as account_views
 from django.contrib.messages.views import SuccessMessageMixin
 from django.core.urlresolvers import reverse
-from django.shortcuts import get_object_or_404
 from django.utils.translation import ugettext as _
 from django.views import generic
 
@@ -12,7 +10,6 @@ from adhocracy4.projects import models as project_models
 from adhocracy4.rules import mixins as rules_mixins
 
 from apps.organisations.models import Organisation
-from apps.users.models import User
 
 from . import mixins as dashboard_mixins
 from . import blueprints
@@ -29,6 +26,7 @@ class DashboardProjectListView(dashboard_mixins.DashboardBaseMixin,
     filter_set = DashboardProjectFilterSet
     template_name = 'meinberlin_dashboard/project_list.html'
     permission_required = 'meinberlin_organisations.initiate_project'
+    menu_item = 'project'
 
     def get_queryset(self):
         return super().get_queryset().filter(
@@ -42,6 +40,7 @@ class DashboardBlueprintListView(dashboard_mixins.DashboardBaseMixin,
     template_name = 'meinberlin_dashboard/blueprint_list.html'
     blueprints = blueprints.blueprints
     permission_required = 'meinberlin_organisations.initiate_project'
+    menu_item = 'project'
 
 
 class DashboardProjectCreateView(dashboard_mixins.DashboardBaseMixin,
@@ -54,6 +53,7 @@ class DashboardProjectCreateView(dashboard_mixins.DashboardBaseMixin,
     template_name = 'meinberlin_dashboard/project_create_form.html'
     success_message = _('Project succesfully created.')
     permission_required = 'meinberlin_organisations.initiate_project'
+    menu_item = 'project'
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
@@ -77,6 +77,7 @@ class DashboardProjectUpdateView(dashboard_mixins.DashboardBaseMixin,
     template_name = 'meinberlin_dashboard/project_update_form.html'
     success_message = _('Project successfully updated.')
     permission_required = 'meinberlin_organisations.initiate_project'
+    menu_item = 'project'
 
     def get_queryset(self):
         return super().get_queryset().filter(
@@ -116,35 +117,7 @@ class DashboardOrganisationUpdateView(dashboard_mixins.DashboardBaseMixin,
     template_name = 'meinberlin_dashboard/organisation_form.html'
     success_message = _('Organisation successfully updated.')
     permission_required = 'meinberlin_organisations.modify_organisation'
-
-
-class DashboardEmailView(dashboard_mixins.DashboardBaseMixin,
-                         account_views.EmailView):
-    menu_item = 'email'
-    template_name = 'meinberlin_dashboard/email.html'
-
-
-class DashboardProfileView(dashboard_mixins.DashboardBaseMixin,
-                           SuccessMessageMixin,
-                           generic.UpdateView):
-
-    model = User
-    template_name = "meinberlin_dashboard/profile.html"
-    form_class = forms.ProfileForm
-    success_message = _("Your profile was successfully updated.")
-    menu_item = 'profile'
-
-    def get_object(self):
-        return get_object_or_404(User, pk=self.request.user.id)
-
-
-class ChangePasswordView(dashboard_mixins.DashboardBaseMixin,
-                         account_views.PasswordChangeView):
-    menu_item = 'password'
-    template_name = 'meinberlin_dashboard/password.html'
-
-    def get_success_url(self):
-        return reverse('dashboard-password')
+    menu_item = 'organisation'
 
 
 class DashboardProjectModeratorsView(dashboard_mixins.DashboardBaseMixin,
@@ -156,6 +129,7 @@ class DashboardProjectModeratorsView(dashboard_mixins.DashboardBaseMixin,
     form_class = forms.AddModeratorForm
     template_name = 'meinberlin_dashboard/project_moderators.html'
     permission_required = 'meinberlin_organisations.initiate_project'
+    menu_item = 'project'
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
