+++ apps/users/templates/meinberlin_users/indicator.html
@@ -17,7 +17,7 @@
     </button>
     <ul class="dropdown-menu" aria-labelledby="user-actions">
         <li>
-            <a class="dropdown-item" href="{% url 'dashboard-profile' %}">{% trans "Account Settings" %}</a>
+            <a class="dropdown-item" href="{% url 'account' %}">{% trans "Account Settings" %}</a>
         </li>
 
         {% for organisation in request.user.organisations %}
